+++ datafusion-expr/Cargo.toml
@@ -37,3 +37,4 @@ path = "src/lib.rs"
 [dependencies]
 datafusion-common = { path = "../datafusion-common", version = "6.0.0" }
 arrow = { version = "8.0.0", features = ["prettyprint"] }
+sqlparser = "0.13"
